+++ TokenService/Models/DataModels/User.cs
@@ -1,13 +1,13 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
 
-namespace TokenService.Models;
+namespace TokenService.Models.DataModels;
 
-public class RefreshTokenModel
+public class User
 {
     [Key]
     public long Id { get; set; }
     public string? UserName { get; set; }
     public string? RefreshToken { get; set; }
     public DateTime RefreshTokenExpiryTime { get; set; }
+    public bool IsRevoked { get; set; }
 }
